% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annot}
\alias{annot}
\title{Genes and associated methylation probes}
\format{
A named list with 5 components - each component corresponds to one gene:
\describe{
each component of the list is a character vector with probe names associated with given gene
}
}
\source{
\url{https://www.cancer.gov/tcga}
}
\usage{
annot
}
\description{
A named list containing the associated methylation probes of given gene.
}
\keyword{datasets}
