% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{B_prior_mat_weighted}
\alias{B_prior_mat_weighted}
\title{Empirical biological knowledge accessor}
\usage{
B_prior_mat_weighted(x)
}
\arguments{
\item{x}{MCMC_sapling_res-class, output from the bn_module function}
}
\value{
Matrix, empirical biological knowledge
}
\description{
\code{B_prior_mat_weighted} This is accessor function for MCMC_sapling_res-class.
}
