% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{beta_tuning}
\alias{beta_tuning}
\title{Beta tuning accessor}
\usage{
beta_tuning(x)
}
\arguments{
\item{x}{MCMC_sapling_res-class, output from the bn_module function}
}
\value{
Matrix, results from adaptive phases that contains hyperparameter
beta tuning
}
\description{
\code{beta_tuning} This is accessor function for MCMC_sapling_res-class.
}
\keyword{internal}
