% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{BGe_node}
\alias{BGe_node}
\title{BGe score for specific node}
\usage{
BGe_node(
  node,
  adjacency_matrix,
  parent_set_combinations,
  BGe_score_all_configs_node
)
}
\arguments{
\item{node}{character vector with given node name.}

\item{adjacency_matrix}{adjacency matrix of given network.}

\item{parent_set_combinations}{list of all possible parent set configuration
for all nodes available.}

\item{BGe_score_all_configs_node}{list of nodes BGe score for all possible
parent set configurations.}
}
\value{
Numeric vector of length 1: BGe score of given node
}
\description{
\code{BGe_node} Computes the BGe score of given node using precomputed sets
of all possible parents.
}
