% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{BGe_score}
\alias{BGe_score}
\title{BGe score}
\usage{
BGe_score(
  adjacency_matrix,
  omics,
  layers_def,
  parent_set_combinations,
  BGe_score_all_configs_node
)
}
\arguments{
\item{adjacency_matrix}{adjacency matrix of given network.}

\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data).
Each component of the list is a matrix with samples in rows and
features in columns.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{parent_set_combinations}{list of all possible parent set configuration
for all nodes available.}

\item{BGe_score_all_configs_node}{list of nodes BGe score for all possible
parent set configurations.}
}
\value{
Numeric vector of length 1: BGe score of given adjacency matrix
}
\description{
\code{BGe_score} Computes the BGe score of given network using precomputed sets
of possible parents.
}
\examples{
data(list=c("PK", "TFtarg_mat", "annot", "layers_def", "omics"),
    package="IntOMICS")
OMICS_mod_res <- OMICS_module(omics = omics, PK = PK, lm_METH = TRUE,
    layers_def = layers_def, TFtargs = TFtarg_mat, annot = annot, 
    r_squared_thres = 0.3)
first.adapt.phase_net <- first_adapt_phase(omics = OMICS_mod_res$omics, 
    B_prior_mat = OMICS_mod_res$B_prior_mat, annot = OMICS_mod_res$annot,
    energy_all_configs_node = OMICS_mod_res$pf_UB_BGe_pre$energy_all_configs_node,
    len = 5, layers_def = OMICS_mod_res$layers_def, prob_mbr = 0.07,
    BGe_score_all_configs_node = OMICS_mod_res$pf_UB_BGe_pre$BGe_score_all_configs_node, 
    parent_set_combinations = OMICS_mod_res$pf_UB_BGe_pre$parents_set_combinations)
adjacency_matrix <- OMICS_mod_res$B_prior_mat
adjacency_matrix[,] <- 0
BGe_score(adjacency_matrix = adjacency_matrix, omics = OMICS_mod_res$omics,
    layers_def = OMICS_mod_res$layers_def, 
    parent_set_combinations =
    OMICS_mod_res$pf_UB_BGe_pre$parents_set_combinations, 
    BGe_score_all_configs_node = OMICS_mod_res$pf_UB_BGe_pre$BGe_score_all_configs_node)

}
