% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BN_mod_res}
\alias{BN_mod_res}
\title{IntOMICS MCMC simulation result}
\format{
A named list with 3 components - each component corresponds to one omics data:
\describe{
\item{B_prior_mat_weighted}{IntOMICS estimated empirical biological knowledge}
\item{sampling.phase_res}{results from the conventional MCMC sampling - two independent simulations}
\item{beta_tuning}{result from the automatically tuned MCMC algorithm}
}
}
\usage{
BN_mod_res
}
\description{
The output from IntOMICS::BN_module function. A named list containing results from the MCMC sampling (resulting sample is thinned and converted into corresponding CPDAGs)
}
\keyword{datasets}
