% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{BN_module}
\alias{BN_module}
\title{#' BN module}
\usage{
BN_module(burn_in, thin, OMICS_mod_res, minseglen, len = 5, prob_mbr = 0.07)
}
\arguments{
\item{burn_in}{numeric vector the minimal length of burn-in period
of the MCMC simulation.}

\item{thin}{numeric vector thinning frequency of the resulting MCMC
simulation.}

\item{OMICS_mod_res}{list output from the OMICS_module function.}

\item{minseglen}{numeric vector minimal number of iterations
with the c_rms value below the c_rms threshold.}

\item{len}{numeric vector initial width of the sampling interval
for hyperparameter beta.}

\item{prob_mbr}{numeric vector probability of the MBR step.}
}
\value{
Large List of 3 elements: empirical biological matrix,
sampling phase result and hyperparameter beta tuning trace
}
\description{
\code{BN_module} Performs automatically tuned MCMC sampling from posterior
distribution together with conventional MCMC sampling using empirical
biological prior matrix to sample network structures from posterior
distribution.
}
\examples{
data(list=c("PK", "TFtarg_mat", "annot", "gene_annot", "layers_def", 
    "omics"), package="IntOMICS")
OMICS_mod_res <- OMICS_module(omics = omics, PK = PK, 
    layers_def = layers_def, TFtargs = TFtarg_mat,
    annot = annot, r_squared_thres = 0.3, 
    lm_METH = TRUE)
BN_mod_res <- BN_module(burn_in = 100000, thin = 500, 
    OMICS_mod_res = OMICS_mod_res, minseglen = 50000, 
    len = 5, prob_mbr = 0.07)

}
