% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borders_def.R
\name{borders_def}
\alias{borders_def}
\title{Color scales}
\usage{
borders_def(node_list, layers_def, omics, omics_meth_original)
}
\arguments{
\item{node_list}{character vector indicating the complete set of nodes
in the resulting network structure.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data).}

\item{omics_meth_original}{METH matrix containing original beta values
Each component of the list is a matrix with samples in rows and features
in columns.}
}
\value{
List of 5 elements indicating the color scale for each modality
}
\description{
\code{borders_def} Determines the color scale for each modality.
}
\keyword{internal}
