% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{CPDAGs_sim1}
\alias{CPDAGs_sim1}
\title{CPDAGs from the first simulation accessor}
\usage{
CPDAGs_sim1(x)
}
\arguments{
\item{x}{MCMC_sapling_res-class, output from the bn_module function}
}
\value{
List, CPDAGs from the first independent MCMC simulation
}
\description{
\code{CPDAGs_sim1} This is accessor function for MCMC_sapling_res-class.
}
\examples{
if(interactive()){data("BN_mod_res", package="IntOMICS_R")
CPDAGs_sim1(BN_mod_res)}

}
