% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{CPDAGs_sim2}
\alias{CPDAGs_sim2}
\title{CPDAGs from the second simulation accessor}
\usage{
CPDAGs_sim2(x)
}
\arguments{
\item{x}{MCMC_sapling_res-class, output from the bn_module function}
}
\value{
List, CPDAGs from the second independent MCMC simulation
}
\description{
\code{CPDAGs_sim2} This is accessor function for MCMC_sapling_res-class.
}
