% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_core_score.R
\name{dag_core_score}
\alias{dag_core_score}
\title{BGe score}
\usage{
dag_core_score(j, parentnodes, n, param)
}
\arguments{
\item{j}{character vector a node to be scored}

\item{parentnodes}{character vector the parents of the node j}

\item{n}{numeric vector number of nodes in the netwrok}

\item{param}{an object of class scoreparameters, which includes all
necessary information for calculating the BDe/BGe score}
}
\value{
Numeric vector of length 1
}
\description{
\code{dag_core_score} The log of the BGe score simplified as much as possible.
This function is from BiDAG package.
}
\keyword{internal}
