% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAGcorescore.R
\name{DAGcorescore}
\alias{DAGcorescore}
\title{BGe score}
\usage{
DAGcorescore(j, parentnodes, n, param)
}
\arguments{
\item{j}{character vector a node to be scored}

\item{parentnodes}{character vector the parents of the node j}

\item{n}{numeric vector number of nodes in the netwrok}

\item{param}{an object of class scoreparameters, which includes all
necessary information for calculating the BDe/BGe score}
}
\value{
Numeric vector of length 1
}
\description{
\code{DAGcorescore} The log of the BGe score simplified as much as possible.
This function is from BiDAG package.
}
