% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens_edge_weights.R
\name{dens_edge_weights}
\alias{dens_edge_weights}
\title{Density plot of edge weights inferred by IntOMICS}
\usage{
dens_edge_weights(net)
}
\arguments{
\item{net}{list output from the weighted_net function.}
}
\value{
density plot of edge weights
}
\description{
\code{dens_edge_weights} Creates density plot of edge weights.
}
\examples{
if(interactive()){data(list=c("OMICS_mod_res", "BN_mod_res", "gene_annot", "TFtarg_mat", 
"PK"), package="IntOMICS")
res_weighted <- edge_weights(mcmc_res = BN_mod_res, burn_in = 10000, 
 thin = 500, edge_freq_thres = 0.3) 
weighted_net_res <- weighted_net(cpdag_weights = res_weighted, 
 gene_annot = gene_annot, PK = PK, OMICS_mod_res = OMICS_mod_res, 
 gene_ID = "gene_symbol", TFtargs = TFtarg_mat,
 B_prior_mat_weighted = B_prior_mat_weighted(BN_mod_res))
dens_edge_weights(weighted_net_res)
}

}
