% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_weights.R
\name{edge_weights}
\alias{edge_weights}
\title{Edge weights of MCMC simulation}
\usage{
edge_weights(mcmc_res, burn_in, thin, edge_freq_thres = NULL)
}
\arguments{
\item{mcmc_res}{list output from the bn_module function.}

\item{burn_in}{numeric vector the minimal length of burn-in period
of the MCMC simulation.}

\item{thin}{numeric vector thinning frequency of the resulting
MCMC simulation.}

\item{edge_freq_thres}{numerical vector the quantile of all edge weights
used to filter the most reliable edges.}
}
\value{
data.frame with edges and corresponding edge weights;
edge_freq_thres used to filter relevant edges
}
\description{
\code{edge_weights} Returns list of edges with corresponding posterior
probabilities (possibly filtered low reliable edges).
}
\examples{
data("BN_mod_res", package="IntOMICS")
res_weighted <- trace_plots(mcmc_res = BN_mod_res, burn_in = 10000, 
       thin = 500, edge_freq_thres = 0.3) 

}
