% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_b_heatmap.R
\name{emp_b_heatmap}
\alias{emp_b_heatmap}
\title{Heatmap of empB - B}
\usage{
emp_b_heatmap(mcmc_res, OMICS_mod_res, gene_annot, TFtargs)
}
\arguments{
\item{mcmc_res}{MCMC_sapling_res output from the bn_module function.}

\item{OMICS_mod_res}{list output from the omics_module function.}

\item{gene_annot}{data.frame containing the entrez ID and corresponding gene
symbol for conversion.}

\item{TFtargs}{matrix containing the direct interactions between TFs
(columns) and their targets (rows).}
}
\value{
Figure heatmap
}
\description{
\code{emp_b_heatmap} plot a heatmap with empB - B values (depicts the difference
between prior knowledge and the empirical knowledge)
}
\examples{
data(list=c("TFtarg_mat", "gene_annot", "OMICS_mod_res",
"BN_mod_res"), package="IntOMICS")
emp_b_heatmap(mcmc_res = BN_mod_res, OMICS_mod_res = OMICS_mod_res, 
    gene_annot = gene_annot, TFtargs = TFtarg_mat)
            
}
