% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_func.R
\name{empB_heatmap}
\alias{empB_heatmap}
\title{Heatmap of empB - B}
\usage{
empB_heatmap(mcmc_res, OMICS_mod_res, gene_annot, TFtargs)
}
\arguments{
\item{mcmc_res}{list output from the BN_module function.}

\item{OMICS_mod_res}{list output from the OMICS_module function.}

\item{gene_annot}{data.frame containing the entrez ID and corresponding gene
symbol for conversion.}

\item{TFtargs}{matrix containing the direct interactions between TFs
(columns) and their targets (rows).}
}
\value{
Figure heatmap
}
\description{
\code{empB_heatmap} plot a heatmap with empB - B values (depicts the difference
between prior knowledge and the empirical knowledge)
}
\examples{
data(list=c("PK", "TFtarg_mat", "annot", "gene_annot", "layers_def",
"omics"), package="IntOMICS")
OMICS_mod_res <- OMICS_module(omics = omics, PK = PK, annot = annot, 
    layers_def = layers_def, TFtargs = TFtarg_mat, r_squared_thres = 0.3, 
    lm_METH = TRUE, gene_annot = gene_annot)
BN_mod_res <- BN_module(burn_in = 100000, thin = 500, len = 5,
    OMICS_mod_res = OMICS_mod_res, minseglen = 50000, prob_mbr = 0.07)
empB_heatmap(mcmc_res = BN_mod_res, OMICS_mod_res = OMICS_mod_res, 
    gene_annot = gene_annot, TFtargs = TFtarg_mat)
            
}
