% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_function_node_specific.R
\name{energy_function_node_specific}
\alias{energy_function_node_specific}
\title{Node energy function}
\usage{
energy_function_node_specific(all_parents_config, B_prior_mat, int_node)
}
\arguments{
\item{all_parents_config}{matrix with all possible parent set configurations
(column indicates parents of given int_node).}

\item{B_prior_mat}{a biological prior matrix.}

\item{int_node}{character vector with given node name.}
}
\value{
Numeric vector of length 1
}
\description{
\code{energy_function_node_specific}  For each node returns its energy over all
parent set configurations, the empty parent set is included.
}
\examples{
data(list=c("PK", "TFtarg_mat", "annot", "layers_def", "omics", "gene_annot"),
package="IntOMICS")
colnames(omics$ge) <- gene_annot$entrezID[match(colnames(omics$ge),
gene_annot$gene_symbol)]
colnames(omics$cnv) <- tolower(gene_annot$entrezID[match(colnames(omics$cnv),
gene_annot$gene_symbol)])
B <- B_prior_mat(omics = omics, PK = PK, layers_def = layers_def, 
   annot = annot, lm_METH = TRUE, r_squared_thres = 0.3, p_val_thres = 0.05,
   TFtargs = TFtarg_mat, TFBS_belief = 0.75, nonGE_belief = 0.5, 
   woPKGE_belief = 0.5)
all_parents_config <- matrix(c("EID:2535", "EID:2535", 
"EID:1857", "EID:2932"), byrow = TRUE, nrow=2)
energy_function_node_specific(all_parents_config = all_parents_config,
B_prior_mat = B$B_prior_mat, int_node = "EID:7482")

}
