% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{estimated_beta}
\alias{estimated_beta}
\title{Estimated beta accessor}
\usage{
estimated_beta(x)
}
\arguments{
\item{x}{MCMC_sapling_res-class, output from the bn_module function}
}
\value{
Numeric, trace of root mean square used for c_rms measure
}
\description{
\code{estimated_beta} This is accessor function for MCMC_sapling_res-class.
}
\examples{
data("BN_mod_res", package="IntOMICS")
estimated_beta(BN_mod_res)

}
