% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{estimated_len}
\alias{estimated_len}
\title{Estimated len accessor}
\usage{
estimated_len(x)
}
\arguments{
\item{x}{MCMC_sapling_res-class, output from the bn_module function}
}
\value{
Numeric, width of the sampling interval for hyperparameter beta
}
\description{
\code{estimated_len} This is accessor function for MCMC_sapling_res-class.
}
