% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{first.adapt.phase_net}
\alias{first.adapt.phase_net}
\title{IntOMICS first adaption phase result}
\format{
A named list with 10 components:
\describe{
\item{source.net}{initial adjacency matrix}
\item{beta.source}{initial beta value}
\item{partition_func_UB_beta_source}{partition function upper bound}
\item{acceptance_saved}{acceptance ratio}
\item{B_prior_mat}{biological prior matrix}
\item{acceptance_beta_saved}{acceptance ratio of beta value}
\item{betas}{simulated beta values}
\item{method_choice_saved}{MCMC method used to sample network structure}
\item{nets}{simulated networks}
\item{energy_all_configs_node}{energy for all possible parent set configurations}
}
}
\usage{
first.adapt.phase_net
}
\description{
The output from IntOMICS::first_adapt_phase function. A named list containing results from the MCMC sampling of the first adaption phase.
}
\keyword{datasets}
