% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_annot}
\alias{gene_annot}
\title{Gene ID conversion table}
\format{
A data.frame with 8 rows and 2 variables:
\describe{
\item{entrezID}{Entrez ID}
\item{gene_symbol}{gene symbol}
}
}
\usage{
gene_annot
}
\description{
A data.frame containing the entrez ID and corresponding gene symbol.
}
\keyword{datasets}
