% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{init.net.mcmc}
\alias{init.net.mcmc}
\title{Random initial network}
\usage{
init.net.mcmc(omics, layers_def, B_prior_mat)
}
\arguments{
\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data). Each component of the list is a matrix
with samples in rows and features in columns.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{B_prior_mat}{a biological prior matrix.}
}
\value{
List of 2 elements: random adjacency network and empty network
}
\description{
\code{init.net.mcmc} This function is used to sample random initial network.
The edges are sampled only between GE nodes.
}
