% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntOMICSr-package.R
\docType{package}
\name{IntOMICSr-package}
\alias{IntOMICSr}
\alias{IntOMICSr-package}
\title{IntOMICSr: Integrative analysis of multi-omics data to infer regulatory networks}
\description{
IntOMICSr is an efficient integrative framework based on Bayesian networks. IntOMICSr systematically analyses gene expression (GE), DNA methylation (METH), copy number variation (CNV) and biological prior knowledge (B) to infer regulatory networks. IntOMICSr complements the missing biological prior knowledge by so-called empirical biological knowledge (empB), estimated from the available experimental data. An automatically tuned MCMC algorithm (Yang and Rosenthal, 2017) estimates model parameters and the empirical biological knowledge. Conventional MCMC algorithm with additional Markov blanket resampling (MBR) step (Su and Borsuk, 2016) infers resulting regulatory network structure consisting of three types of nodes: GE nodes refer to gene expression levels, CNV nodes refer to associated copy number variations, and METH nodes refer to associated DNA methylation probe(s).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anna-pacinkova/IntOMICSr}
  \item Report bugs at \url{https://github.com/anna-pacinkova/IntOMICSr/issues}
}

}
\author{
\strong{Maintainer}: Pacinkova Anna \email{ana.pacinkova@gmail.com}

}
\keyword{internal}
