% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_acyclic.R
\name{is_acyclic}
\alias{is_acyclic}
\title{Acyclic network identification.}
\usage{
is_acyclic(g)
}
\arguments{
\item{g}{adajcency matrix of given network/graph.}
}
\value{
boolean of length 1
}
\description{
\code{is_acyclic} This function is from bnstruct R package. Check if the directed
graph is acyclic.
}
\examples{
is_acyclic(matrix(c(1,rep(0,20)), nrow=3))

}
\keyword{internal}
