% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layers_def}
\alias{layers_def}
\title{Gene ID conversion table}
\format{
A data.frame with 3 rows and 3 variables:
\describe{
\item{omics}{modality}
\item{layer}{layer ID}
\item{fan_in_ge}{maximal number of parents from given layer to single GE node}
}
}
\usage{
layers_def
}
\description{
A data.frame containing the modality ID, corresponding layer in BN and maximal number of parents from given layer to GE nodes.
}
\keyword{datasets}
