% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_custom_ggplot.R
\name{legend_custom_ggplot}
\alias{legend_custom_ggplot}
\title{Node color legend}
\usage{
legend_custom_ggplot(net)
}
\arguments{
\item{net}{list output from the trace_plots function.}
}
\value{
Figure with color key
}
\description{
\code{legend_custom_ggplot} The color scale for each modality.
}
\examples{
data(list=c("OMICS_mod_res", "BN_mod_res", "gene_annot", "TFtarg_mat", 
"PK"), package="IntOMICS")
res_weighted <- edge_weights(mcmc_res = BN_mod_res, burn_in = 10000, 
 thin = 500, edge_freq_thres = 0.3) 
weighted_net_res <- weighted_net(cpdag_weights = res_weighted, 
 gene_annot = gene_annot, PK = PK, OMICS_mod_res = OMICS_mod_res, 
 gene_ID = "gene_symbol", TFtargs = TFtarg_mat,
 B_prior_mat_weighted = BN_mod_res@B_prior_mat_weighted) 
legend_custom_ggplot(weighted_net_res)

}
