% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr.R
\name{mbr}
\alias{mbr}
\title{Markov Blanket Resampling}
\usage{
mbr(
  source_net_adjacency,
  layers_def,
  omics,
  BGe_score_all_configs_node,
  parent_set_combinations
)
}
\arguments{
\item{source_net_adjacency}{adajcency matrix of given network.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data). Each component of the list is a matrix
with samples in rows and features in columns.}

\item{BGe_score_all_configs_node}{list of nodes BGe score for all possible
parent set configurations.}

\item{parent_set_combinations}{list of all possible parent set configuration
for all nodes available.}
}
\value{
List of 10 elements needed to define adjacency matrix
with markov blanket resampling
}
\description{
\code{mbr} This function performs the markov blanket resampling method according
to Su and Borsuk, 2016.
}
\examples{
data("OMICS_mod_res", package="IntOMICS")
mbr(source_net_adjacency = OMICS_mod_res$B_prior_mat, 
    layers_def = OMICS_mod_res$layers_def, omics = OMICS_mod_res$omics, 
    BGe_score_all_configs_node = 
    OMICS_mod_res$pf_UB_BGe_pre$BGe_score_all_configs_node,
    parent_set_combinations = 
    OMICS_mod_res$pf_UB_BGe_pre$parents_set_combinations)

}
\keyword{internal}
