% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_classes.R
\docType{class}
\name{MCMC_sapling_res}
\alias{MCMC_sapling_res}
\alias{MCMC_sapling_res-class}
\title{The MCMC_sapling_res class}
\description{
Container of an MCMC sampling phase results generated
by the function \code{\link[IntOMICS_R]{bn_module}}.
}
\section{Slots}{

\describe{
\item{\code{estimated_beta}}{Numeric, estimated value of
hyperparameter beta}

\item{\code{estimated_len}}{Numeric, estimated width of the sampling
interval for hyperparameter beta}

\item{\code{B_prior_mat_weighted}}{Empirical biological knowledge matrix,
interactions from the biological prior knowledge and TFs-target
interactions are constant (unless if "TFBS_belief" is not equal
to "woPKGE_belief").}

\item{\code{CPDAGs_sim1}}{List of CPDAGs from the first independent MCMC
simulation (thinned DAGs from the MCMC simulation converted
into CPDAGs, ducplicated CPDAGs discarded)}

\item{\code{CPDAGs_sim2}}{List of CPDAGs from the second independent MCMC
simulation (thinned DAGs from the MCMC simulation converted
into CPDAGs, ducplicated CPDAGs discarded)}

\item{\code{beta_tuning}}{Matrix of results from adaptive phases that
contains hyperparameter beta tuning
\describe{
\item{\code{value}}{trace of hyperparameter beta}
\item{len}{trace of width of the sampling interval for hyperparameter
beta}}}

\item{\code{rms}}{Numeric, trace of root mean square used for c_rms
measure to evaluate the convergence of MCMC simulation}
}}

\examples{

# A MCMC_sapling_res object created by the bn_module function.
if(interactive()){data("OMICS_mod_res", package="IntOMICS_R")
BN_mod_res <- bn_module(burn_in = 500, 
    thin = 20, OMICS_mod_res = OMICS_mod_res, 
    minseglen = 5, len = 5, prob_mbr = 0.07)}

}
