% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_simulation_sampling_phase.R
\name{mcmc_simulation_sampling_phase}
\alias{mcmc_simulation_sampling_phase}
\title{Sampling phase}
\usage{
mcmc_simulation_sampling_phase(
  first,
  last,
  sim_init,
  prob_mbr,
  B_prior_mat,
  omics,
  parent_set_combinations,
  BGe_score_all_configs_node,
  layers_def,
  len,
  thin,
  energy_all_configs_node,
  annot
)
}
\arguments{
\item{first}{numeric vector iteration to start.}

\item{last}{numeric vector iteration to stop.}

\item{sim_init}{list output from the source_net_def function or from two
independent simulations from the mcmc_simulation_sampling_phase function.}

\item{prob_mbr}{numeric vector probability of the MBR step.}

\item{B_prior_mat}{a biological prior matrix.}

\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data).
Each component of the list is a matrix with samples in rows and features
in columns.}

\item{parent_set_combinations}{list of all possible parent set configuration
for all nodes available.}

\item{BGe_score_all_configs_node}{list of nodes BGe score for all possible
parent set configurations.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{len}{numeric vector initial width of the sampling interval
for hyperparameter beta.}

\item{thin}{numeric vector thinning frequency of the resulting
MCMC simulation.}

\item{energy_all_configs_node}{list of nodes energy for all possible parent
set configurations.}

\item{annot}{named list containing the associated methylation probes
of given gene.}
}
\value{
List of 1 element: sampling phase result before MCMC convergence
}
\description{
\code{mcmc_simulation_sampling_phase} This function performs the final sampling
of network structures with estimated hyperparameters.
It if part of sampling_phase function.
}
\keyword{internal}
