% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{neighborhood_size}
\alias{neighborhood_size}
\title{Neighborhood size}
\usage{
neighborhood_size(net, layers_def, B_prior_mat, omics)
}
\arguments{
\item{net}{adajcency matrix of given network.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{B_prior_mat}{a biological prior matrix.}

\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data). Each component of the list is a matrix
with samples in rows and features in columns.}
}
\value{
Numeric of length 1: neighborhood size
}
\description{
\code{neighborhood_size} This function is determines number of network structures
that can be reached from the current network structure.
}
\examples{
data(list=c("PK", "TFtarg_mat", "annot", "layers_def", "omics", "gene_annot"),
package="IntOMICS")
colnames(omics$ge) <- gene_annot$entrezID[match(colnames(omics$ge),
gene_annot$gene_symbol)]
colnames(omics$cnv) <- tolower(gene_annot$entrezID[match(colnames(omics$cnv),
gene_annot$gene_symbol)])
B <- B_prior_mat(omics = omics, PK = PK, layers_def = layers_def, 
     annot = annot, lm_METH = TRUE, r_squared_thres = 0.3,
     p_val_thres = 0.05, TFtargs = TFtarg_mat, TFBS_belief = 0.75, 
     nonGE_belief = 0.5, woPKGE_belief = 0.5)
adjacency_matrix <- B$B_prior_mat
adjacency_matrix[,] <- 0
neighborhood_size(net = adjacency_matrix, layers_def = layers_def,
B_prior_mat = B$B_prior_mat, omics = B$omics)

}
