% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omics}
\alias{omics}
\title{Omics data}
\format{
A MatchedAssayExperiment with 3 components - each component corresponds to
one omics data:
\describe{
MatchedAssayExperiment with variable number of columns
}
}
\source{
\url{https://www.cancer.gov/tcga}
}
\usage{
omics
}
\description{
A MatchedAssayExperiment with names same as in layers_def$omics column
containing the gene expression, copy number variation and methylation data.
}
\keyword{datasets}
