% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OMICS_mod_res}
\alias{OMICS_mod_res}
\title{preprocessed IntOMICSr input data}
\format{
A named list with 6 components:
\describe{
\item{pf_UB_BGe_pre}{output from IntOMICSr::pf_UB_est function}
\item{B_prior_mat}{biological prior matrix}
\item{annot}{genes and associated methylation probes}
\item{omics}{a named list containing the gene expression, copy number variation and methylation data}
\item{layers_def}{layers definition of all omics data}
\item{omics_meth_original}{original methylation data}
}
}
\usage{
OMICS_mod_res
}
\description{
The output from IntOMICSr::OMICS_module function. A named list containing preprocessed input data.
}
\keyword{datasets}
