% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{OMICS_module}
\alias{OMICS_module}
\title{OMICS_module}
\usage{
OMICS_module(
  omics,
  PK = NULL,
  layers_def,
  TFtargs = NULL,
  annot = NULL,
  lm_METH = TRUE,
  r_squared_thres = 0.3,
  p_val_thres = 0.05,
  TFBS_belief = 0.75,
  nonGE_belief = 0.5,
  woPKGE_belief = 0.5,
  gene_annot
)
}
\arguments{
\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data).
Each component of the list is a matrix with samples in rows and
features in columns.}

\item{PK}{data.frame with known interactions.}

\item{layers_def}{data.frame containing the modality ID, corresponding
layer in BN and maximal number of parents from given layer to GE nodes.}

\item{TFtargs}{matrix containing the direct interactions between TFs
(columns) and their targets (rows).}

\item{annot}{named list containing the associated methylation probes
of given gene.}

\item{lm_METH}{logical asking whether to use linear regression to filter
methylation data (default=TRUE).}

\item{r_squared_thres}{numeric vector to define the R^2 used as a threshold
of significance in linear regression if lm_METH=TRUE (default=0.3).}

\item{p_val_thres}{numeric vector to define the p-value used as a threshold
of significance in linear regression if lm_METH=TRUE (default=0.05).}

\item{TFBS_belief}{numeric vector to define the belief concerning the TF
and its target interaction (default=0.75).}

\item{nonGE_belief}{numeric vector to define the belief concerning
interactions of features except GE-GE (default=0.5).}

\item{woPKGE_belief}{numeric vector to define the belief concerning GE-GE
interactions without prior knowledge (default=0.5).}

\item{gene_annot}{data.frame containing the entrez ID and corresponding
gene symbol for conversion.}
}
\value{
List of 6 elements needed to init MCMC simulation
}
\description{
\code{OMICS_module} data preprocessing + B_prior_mat definition +
partition function upper bound estimation +
all possible parent sets per node definition +
BGe score computation for all possible parent sets
}
\examples{
data(list=c("PK", "TFtarg_mat", "annot", "layers_def", "omics", 
"gene_annot"), package="IntOMICS")
OMICS_mod_res <- OMICS_module(omics = omics, PK = PK, 
    layers_def = layers_def, TFtargs = TFtarg_mat, annot = annot, 
    gene_annot = gene_annot, r_squared_thres = 0.3, lm_METH = TRUE)

}
