% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_to_list.R
\name{omics_to_list}
\alias{omics_to_list}
\title{Convert omics MatchedAssayExperiment to list}
\usage{
omics_to_list(omics, layers_def, gene_annot)
}
\arguments{
\item{omics}{MatchedAssayExperiment containing the gene expression
(possibly copy number variation and methylation data).}

\item{layers_def}{data.frame containing the modality ID, corresponding
layer in BN and maximal number of parents from given layer to GE nodes.}

\item{gene_annot}{data.frame containing the entrez ID and corresponding
gene symbol for conversion.}
}
\value{
List of omics modalities
}
\description{
\code{omics_to_list} converts omics MatchedAssayExperiment to list
}
\examples{
data(list=c("layers_def", "omics", "gene_annot"), package="IntOMICS")
omics_to_list(omics = omics,layers_def = layers_def,
gene_annot = gene_annot)

}
