% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_sets_sum_scores_childrenX.R
\name{parent_sets_sum_scores_childrenX}
\alias{parent_sets_sum_scores_childrenX}
\title{MBR sum of children scores}
\usage{
parent_sets_sum_scores_childrenX(
  parent_set_combinations,
  selected_node,
  children_selected_node,
  child_order,
  dag_tmp_bn,
  new_parent_set,
  source_net_adjacency,
  BGe_score_all_configs_node
)
}
\arguments{
\item{parent_set_combinations}{list of all possible parent set configuration
for all nodes available.}

\item{selected_node}{character vector with given node name.}

\item{children_selected_node}{character vector with children
of selected_node in given network structure.}

\item{child_order}{numeric vector random order of children_selected_node.}

\item{dag_tmp_bn}{object of class bn reflecting given network structure.}

\item{new_parent_set}{logical asking whether to define new parent set
for selected_node.}

\item{source_net_adjacency}{adajcency matrix of given network.}

\item{BGe_score_all_configs_node}{list of nodes BGe score for all possible
parent set configurations.}
}
\value{
List of 3 elements
}
\description{
\code{parent_sets_sum_scores_childrenX} This function determines the sum of BGe
scores of given node's children.
}
