% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{pf_UB_est}
\alias{pf_UB_est}
\title{Partition function upper bound}
\usage{
pf_UB_est(omics, B_prior_mat, layers_def, annot)
}
\arguments{
\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data). Each component of the list is a matrix
with samples in rows and features in columns.}

\item{B_prior_mat}{a biological prior matrix.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{annot}{named list containing the associated methylation probes
of given gene.}
}
\value{
List of 4 elements needed to simulate MCMC sampling
}
\description{
\code{pf_UB_est} Partition function upper bound estimation with beta = 0.
For each node returns energy over all possible parent set configurations
and BGe score.
}
