% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PK}
\alias{PK}
\title{Wnt signalling pathway}
\format{
A data.frame with 6 rows and 3 variables:
\describe{
\item{src_entrez}{the parent node}
\item{dest_entrez}{the child node}
\item{edge_type}{the edge from parent node to child node is present or missing}
}
}
\source{
\url{https://www.kegg.jp/entry/map04310}
}
\usage{
PK
}
\description{
A dataset containing known direct interactions between 7 genes.
}
\keyword{datasets}
