% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_01.R
\name{range_01}
\alias{range_01}
\title{Range between 0 and 1}
\usage{
range_01(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Numeric vector with normalised values
}
\description{
\code{range_01} This function re-scales a numeric vector so that it ranges
between 0 and 1.
}
\examples{
range_01(stats::rnorm(10))

}
\keyword{internal}
