% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{rms}
\alias{rms}
\title{c_rms trace accessor}
\usage{
rms(x)
}
\arguments{
\item{x}{MCMC_sapling_res-class, output from the bn_module function}
}
\value{
Numeric, trace of root mean square used for c_rms measure
}
\description{
\code{rms} This is accessor function for MCMC_sapling_res-class.
}
\keyword{internal}
