% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chain.R
\name{sample_chain}
\alias{sample_chain}
\title{Random initial network edge generation}
\usage{
sample_chain(empty_net, omics_ge)
}
\arguments{
\item{empty_net}{adjacency matrix of an empty network/graph
(all values are 0).}

\item{omics_ge}{matrix with gene expression data (samples in rows and
features in columns).}
}
\value{
BN object with conditional probabilities
}
\description{
\code{sample_chain} This function is used to sample random initial network.
The edges are sampled only between GE nodes.
}
