% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_parameters_bidag_bge.R
\name{score_parameters_bidag_bge}
\alias{score_parameters_bidag_bge}
\title{BGe score parameters}
\usage{
score_parameters_bidag_bge(n, data, bgepar = list(am = 1, aw = NULL))
}
\arguments{
\item{n}{numeric number of columns in data matrix.}

\item{data}{matrix with features in columns and a number of rows equal
to the number of samples.}

\item{bgepar}{list which contains parameters for BGe score computation.}
}
\value{
Object of class scoreparameters, which includes all necessary
information for calculating the BDe/BGe score
}
\description{
\code{score_parameters_bidag_bge} Returns parameters needed for calculation
of the BGe score. This function is from BiDAG package.
}
\keyword{internal}
