% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreparameters_BiDAG_BGe.R
\name{scoreparameters_BiDAG_BGe}
\alias{scoreparameters_BiDAG_BGe}
\title{BGe score parameters}
\usage{
scoreparameters_BiDAG_BGe(n, data, bgepar = list(am = 1, aw = NULL))
}
\arguments{
\item{n}{numeric number of columns in data matrix.}

\item{data}{matrix with features in columns and a number of rows equal
to the number of samples.}

\item{bgepar}{list which contains parameters for BGe score computation.}
}
\value{
Object of class scoreparameters, which includes all necessary
information for calculating the BDe/BGe score
}
\description{
\code{scoreparameters_BiDAG_BGe} Returns parameters needed for calculation
of the BGe score. This function is from BiDAG package.
}
