% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_classes.R
\name{show,MCMC_sapling_res-method}
\alias{show,MCMC_sapling_res-method}
\title{MCMC_sampling_res-methods}
\usage{
\S4method{show}{MCMC_sapling_res}(object)
}
\arguments{
\item{object}{gien MCMC_sampling_res-class object}
}
\value{
Get summary of the properties of MCMC_sampling_res-class object.
}
\description{
set show method for MCMC_sampling_res-class objects.
}
\keyword{internal}
