% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NET_manipulation_func.R
\name{source_net_def}
\alias{source_net_def}
\title{Source network for MCMC simulation}
\usage{
source_net_def(
  init.net.mcmc.output,
  parent_set_combinations,
  omics,
  BGe_score_all_configs_node,
  B_prior_mat,
  layers_def,
  energy_all_configs_node,
  len
)
}
\arguments{
\item{init.net.mcmc.output}{list output of the init.net.mcmc function.}

\item{parent_set_combinations}{list of all possible parent set
configuration for all nodes available.}

\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data). Each component of the list is a matrix
with samples in rows and features in columns.}

\item{BGe_score_all_configs_node}{list of nodes BGe score for all possible
parent set configurations.}

\item{B_prior_mat}{a biological prior matrix.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{energy_all_configs_node}{list of nodes energy for all possible parent
set configurations.}

\item{len}{numeric vector initial width of the sampling interval
for hyperparameter beta.}
}
\value{
List of 10 elements needed to define the initial adjacency matrix
}
\description{
\code{source_net_def} This function is used to create the initial network
with its features necessary for MCMC simulation.
}
\examples{
data(list=c("PK", "TFtarg_mat", "annot", "layers_def", "omics"),
package="IntOMICS")
OMICS_mod_res <- OMICS_module(omics = omics, PK = PK, 
    layers_def = layers_def, TFtargs = TFtarg_mat, annot = annot, 
    r_squared_thres = 0.3, lm_METH = TRUE)
init.net <- init.net.mcmc(omics = OMICS_mod_res$omics, 
    layers_def = OMICS_mod_res$layers_def, 
    B_prior_mat = OMICS_mod_res$B_prior_mat)
source_net_def(init.net.mcmc.output = init.net, 
    omics = OMICS_mod_res$omics, 
    parent_set_combinations = OMICS_mod_res$pf_UB_BGe_pre$parents_set_combinations,
    BGe_score_all_configs_node = OMICS_mod_res$pf_UB_BGe_pre$BGe_score_all_configs_node, 
    B_prior_mat = OMICS_mod_res$B_prior_mat, 
    layers_def = OMICS_mod_res$layers_def, len = 5,
    energy_all_configs_node = OMICS_mod_res$pf_UB_BGe_pre$energy_all_configs_node)
             
}
