% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TFtarg_mat}
\alias{TFtarg_mat}
\title{transcription factors and their known targets}
\format{
A matrix with 22452 rows and 181 variables:
\describe{
columns refer to TFs and rows to their targets
}
}
\source{
\url{https://maayanlab.cloud/Harmonizome/dataset/ENCODE+Transcription+Factor+Targets}
}
\usage{
TFtarg_mat
}
\description{
A dataset containing the direct interactions between TFs and their targets.
}
\keyword{datasets}
