% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_net.R
\name{weighted_net}
\alias{weighted_net}
\title{Resulting network definition}
\usage{
weighted_net(
  cpdag_weights,
  gene_annot,
  PK = NULL,
  OMICS_mod_res,
  edge_weights = "MCMC_freq",
  gene_ID,
  TFtargs = NULL,
  B_prior_mat_weighted
)
}
\arguments{
\item{cpdag_weights}{data.frame output from the edge_weights function.}

\item{gene_annot}{data.frame containing the entrez ID and corresponding gene
symbol for conversion.}

\item{PK}{data.frame with known interactions.}

\item{OMICS_mod_res}{list output from the omics_module function.}

\item{edge_weights}{character vector includes either "MCMC_freq" to reflect
the edge weights frequency over the final set of network structures or
"empB" to reflect the empirical biological knowledge estimated by IntOMICS.}

\item{gene_ID}{character vector includes either "gene_symbol" or "entrezID"
to reflect gene identifiers used in the final figure.}

\item{TFtargs}{matrix containing the direct interactions between TFs
(columns) and their targets (rows).}

\item{B_prior_mat_weighted}{matrix one of the outputs
of the bn_module function.}
}
\value{
List of 7 elements needed to plot the final regulatory network
}
\description{
\code{weighted_net} Defines the resulting network structure and determines
the color scale for each modality.
}
\examples{
if(interactive()){data(list=c("OMICS_mod_res", "BN_mod_res", "gene_annot", "TFtarg_mat", 
"PK"), package="IntOMICS")
res_weighted <- edge_weights(mcmc_res = BN_mod_res, burn_in = 10000, 
 thin = 500, edge_freq_thres = 0.3) 
weighted_net_res <- weighted_net(cpdag_weights = res_weighted, 
 gene_annot = gene_annot, PK = PK, OMICS_mod_res = OMICS_mod_res, 
 gene_ID = "gene_symbol", TFtargs = TFtarg_mat,
 B_prior_mat_weighted = B_prior_mat_weighted(BN_mod_res)) 
}

}
