% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptance_check.R
\name{acceptance_check}
\alias{acceptance_check}
\title{Acceptance rate checking}
\usage{
acceptance_check(
  first.adapt.phase_net,
  round_check,
  last_iter_check,
  prob_mbr,
  layers_def,
  parent_set_combinations,
  BGe_score_all_configs_node,
  omics,
  annot
)
}
\arguments{
\item{first.adapt.phase_net}{list output of the first.adapt.phase
or source_net_def function.}

\item{round_check}{numeric vector after each round_check iterations
for which we calculate the beta acceptance rate.}

\item{last_iter_check}{numeric vector number of the acceptance rate
for the past last_iter_check iterations.}

\item{prob_mbr}{numeric vector probability of the MBR step.}

\item{layers_def}{data.frame containing the modality ID, corresponding
layer in BN and maximal number of parents from given layer to GE nodes.}

\item{parent_set_combinations}{list of all possible parent set
configuration for all nodes available.}

\item{BGe_score_all_configs_node}{list of nodes BGe score for all possible
parent set configurations.}

\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data). Each component of the list is a matrix
with samples in rows and features in columns.}

\item{annot}{named list containing the associated methylation
probes of given gene.}
}
\value{
List of 1 element: first adaption phase result
before given acceptance rate
}
\description{
\code{acceptance_check} This phase verify if the acceptance is in range
of 0.28 and 0.6.
}
\keyword{internal}
