% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMICS_module_func.R
\name{B_prior_mat}
\alias{B_prior_mat}
\title{B_prior_mat}
\usage{
B_prior_mat(
  omics,
  PK,
  layers_def,
  TFtargs,
  annot,
  lm_METH,
  r_squared_thres,
  p_val_thres,
  TFBS_belief,
  nonGE_belief,
  woPKGE_belief
)
}
\arguments{
\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data).
Each component of the list is a matrix with samples in rows and features
in columns.}

\item{PK}{data.frame with known interactions.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{TFtargs}{matrix containing the direct interactions between TFs
(columns) and their targets (rows).}

\item{annot}{named list containing the associated methylation probes
of given gene.}

\item{lm_METH}{logical asking whether to use linear regression to filter
methylation data (default=TRUE).}

\item{r_squared_thres}{numeric vector to define the R^2 used as a threshold
of significance in linear regression if lm_METH=TRUE (default=0.3).}

\item{p_val_thres}{numeric vector to define the p-value used as a threshold
of significance in linear regression if lm_METH=TRUE (default=0.05).}

\item{TFBS_belief}{numeric vector to define the belief concerning the TF and
its target interaction (default=0.75).}

\item{nonGE_belief}{numeric vector to define the belief concerning
interactions of features except GE-GE (default=0.5).}

\item{woPKGE_belief}{numeric vector to define the belief concerning GE-GE
interactions without prior knowledge (default=0.5).}
}
\value{
List of 4 elements: prior biological matrix and data
preprocessing
}
\description{
'B_prior_mat' creates the biological prior matrix.
}
