% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bge_node.R
\name{bge_node}
\alias{bge_node}
\title{BGe score for specific node}
\usage{
bge_node(
  node,
  adjacency_matrix,
  parent_set_combinations,
  BGe_score_all_configs_node
)
}
\arguments{
\item{node}{character vector with given node name.}

\item{adjacency_matrix}{adjacency matrix of given network.}

\item{parent_set_combinations}{list of all possible parent set configuration
for all nodes available.}

\item{BGe_score_all_configs_node}{list of nodes BGe score for all possible
parent set configurations.}
}
\value{
Numeric vector of length 1: BGe score of given node
}
\description{
\code{bge_node} Computes the BGe score of given node using precomputed sets
of all possible parents.
}
\examples{
data("OMICS_mod_res", package="IntOMICS")
bge_node(node = "EID:2535", adjacency_matrix = OMICS_mod_res$B_prior_mat,
    parent_set_combinations =
    OMICS_mod_res$pf_UB_BGe_pre$parents_set_combinations, 
    BGe_score_all_configs_node =
    OMICS_mod_res$pf_UB_BGe_pre$BGe_score_all_configs_node)

}
\keyword{internal}
