% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_module_func.R
\name{BGe_score}
\alias{BGe_score}
\title{BGe score}
\usage{
BGe_score(
  adjacency_matrix,
  omics,
  layers_def,
  parent_set_combinations,
  BGe_score_all_configs_node
)
}
\arguments{
\item{adjacency_matrix}{adjacency matrix of given network.}

\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data).
Each component of the list is a matrix with samples in rows and
features in columns.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{parent_set_combinations}{list of all possible parent set configuration
for all nodes available.}

\item{BGe_score_all_configs_node}{list of nodes BGe score for all possible
parent set configurations.}
}
\value{
Numeric vector of length 1: BGe score of given adjacency matrix
}
\description{
\code{BGe_score} Computes the BGe score of given network using precomputed sets
of possible parents.
}
\examples{
data("OMICS_mod_res", package="IntOMICS")
BGe_score(adjacency_matrix = OMICS_mod_res$B_prior_mat,
    layers_def = OMICS_mod_res$layers_def, omics = OMICS_mod_res$omics, 
    parent_set_combinations =
    OMICS_mod_res$pf_UB_BGe_pre$parents_set_combinations, 
    BGe_score_all_configs_node = 
    OMICS_mod_res$pf_UB_BGe_pre$BGe_score_all_configs_node)

}
