% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_module.R
\name{bn_module}
\alias{bn_module}
\title{#' BN module}
\usage{
bn_module(
  burn_in = 1e+05,
  thin = 500,
  OMICS_mod_res,
  minseglen = 50000,
  len = 5,
  prob_mbr = 0.07
)
}
\arguments{
\item{burn_in}{numeric vector the minimal length of burn-in period
of the MCMC simulation.}

\item{thin}{numeric vector thinning frequency of the resulting MCMC
simulation.}

\item{OMICS_mod_res}{list output from the omics_module function.}

\item{minseglen}{numeric vector minimal number of iterations
with the c_rms value below the c_rms threshold.}

\item{len}{numeric vector initial width of the sampling interval
for hyperparameter beta.}

\item{prob_mbr}{numeric vector probability of the MBR step.}
}
\value{
Large List of 3 elements: empirical biological matrix,
sampling phase result and hyperparameter beta tuning trace
}
\description{
\code{bn_module} Performs automatically tuned MCMC sampling from posterior
distribution together with conventional MCMC sampling using empirical
biological prior matrix to sample network structures from posterior
distribution.
}
\examples{
if(interactive()){data("OMICS_mod_res", package="IntOMICSr")
BN_mod_res <- bn_module(burn_in = 500, 
    thin = 20, OMICS_mod_res = OMICS_mod_res, 
    minseglen = 5, len = 5, prob_mbr = 0.07)}

}
