% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_proposal.R
\name{edge_proposal}
\alias{edge_proposal}
\title{Markov Chain conventional single edge proposal move}
\usage{
edge_proposal(net, candidates, layers_def, ge_nodes, omics, B_prior_mat)
}
\arguments{
\item{net}{adajcency matrix of given network.}

\item{candidates}{numeric vector with IDs of potential edge to be changed.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{ge_nodes}{character vector with GE node names}

\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data).
Each component of the list is a matrix with samples in rows and features
in columns.}

\item{B_prior_mat}{a biological prior matrix.}
}
\value{
List of 6 elements needed to define candidates for conventional
single edge proposal move
}
\description{
\code{edge_proposal} This function samples a conventional single edge proposal
moves (identify those edges that are possible to change in given network
structure)
}
\examples{
data(list=c("OMICS_mod_res", "first.adapt.phase_net"),
package="IntOMICS")
source_net <- first.adapt.phase_net$nets[[length(
    first.adapt.phase_net$nets)]]
ge_nodes <- rownames(source_net$adjacency)[regexpr("EID",
    rownames(source_net$adjacency))>0]
vec <- seq_len(length(c(source_net$adjacency)))
vec <- vec[c(OMICS_mod_res$B_prior_mat)>0]
edge_proposal(net = source_net$adjacency, candidates = vec, 
    layers_def = OMICS_mod_res$layers_def, ge_nodes = ge_nodes, 
    omics = OMICS_mod_res$omics, B_prior_mat = OMICS_mod_res$B_prior_mat)
            
}
