% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_func.R
\name{edge_types}
\alias{edge_types}
\title{Resulting network definition}
\usage{
edge_types(
  mcmc_res,
  PK = NULL,
  gene_annot,
  edge_list,
  node_list,
  OMICS_mod_res,
  edge_weights,
  TFtargs = NULL
)
}
\arguments{
\item{mcmc_res}{list output from the BN_module function.}

\item{PK}{data.frame with known interactions.}

\item{gene_annot}{data.frame containing the entrez ID and corresponding
gene symbol for conversion.}

\item{edge_list}{matrix indicating the interaction between nodes,
the first column indicates the source node, the second column indicates
the target node.}

\item{node_list}{character vector indicating the complete set of nodes
in the resulting network structure.}

\item{OMICS_mod_res}{list output from the OMICS_module function.}

\item{edge_weights}{character vector includes either "MCMC_freq" to reflect
the edge weights frequency over the final set of network structures or
"empB" to reflect the empirical biological knowledge estimated by IntOMICS.}

\item{TFtargs}{matrix containing the direct interactions between TFs
(columns) and their targets (rows).}
}
\value{
List of 6 elements needed to plot the final regulatory network edges
}
\description{
\code{edge_types} Defines the resulting network structure and determines
the color scale for each modality. This is part of trace_plots.
}
