% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_function_node_specific.R
\name{energy_function_node_specific}
\alias{energy_function_node_specific}
\title{Node energy function}
\usage{
energy_function_node_specific(all_parents_config, B_prior_mat, int_node)
}
\arguments{
\item{all_parents_config}{matrix with all possible parent set configurations
(column indicates parents of given int_node).}

\item{B_prior_mat}{a biological prior matrix.}

\item{int_node}{character vector with given node name.}
}
\value{
Numeric vector of length 1
}
\description{
\code{energy_function_node_specific}  For each node returns its energy over all
parent set configurations, the empty parent set is included.
}
\keyword{internal}
