% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_func.R
\name{legend_custom}
\alias{legend_custom}
\title{Node color}
\usage{
legend_custom(net)
}
\arguments{
\item{net}{list output from the trace_plots function.}
}
\value{
Figure with color key
}
\description{
\code{legend_custom} Determines the color scale for each modality.
}
\examples{
data(list=c("PK", "TFtarg_mat", "gene_annot",
"OMICS_mod_res", "BN_mod_res"), package="IntOMICS")
res_weighted <- trace_plots(mcmc_res = BN_mod_res, figures_dir = "figures", 
     burn_in = 10000, thin = 500, gene_annot = gene_annot, PK = PK, 
     OMICS_mod_res = OMICS_mod_res, gene_ID = "gene_symbol", 
     edge_freq_thres = 0.3, TFtargs = TFtarg_mat) 
legend_custom(res_weighted)

}
