% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_func.R
\name{legend_custom}
\alias{legend_custom}
\title{Node color}
\usage{
legend_custom(net)
}
\arguments{
\item{net}{list output from the trace_plots function.}
}
\value{
Figure with color key
}
\description{
\code{legend_custom} Determines the color scale for each modality.
}
\examples{
data(list=c("PK", "TFtarg_mat", "annot", "gene_annot", "layers_def",
"omics"), package="IntOMICS")
OMICS_mod_res <- OMICS_module(omics = omics, PK = PK, 
     layers_def = layers_def, TFtargs = TFtarg_mat, annot = annot, 
     r_squared_thres = 0.3, lm_METH = TRUE, gene_annot = gene_annot)
BN_mod_res <- BN_module(burn_in = 100000, thin = 500, len = 5,
     OMICS_mod_res = OMICS_mod_res, minseglen = 50000, prob_mbr = 0.07)
res_weighted <- trace_plots(mcmc_res = BN_mod_res, figures_dir = "figures", 
     burn_in = 100000, thin = 500, gene_annot = gene_annot, PK = PK, 
     OMICS_mod_res = OMICS_mod_res, gene_ID = "gene_symbol", 
     edge_freq_thres = 0.3, TFtargs = TFtarg_mat) 
legend_custom(res_weighted)

}
