% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_custom_ggplot.R
\name{legend_custom_ggplot}
\alias{legend_custom_ggplot}
\title{Node color legend}
\usage{
legend_custom_ggplot(net)
}
\arguments{
\item{net}{list output from the trace_plots function.}
}
\value{
Figure with color key
}
\description{
\code{legend_custom_ggplot} The color scale for each modality.
}
