% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_size.R
\name{neighborhood_size}
\alias{neighborhood_size}
\title{Neighborhood size}
\usage{
neighborhood_size(net, layers_def, B_prior_mat, omics)
}
\arguments{
\item{net}{adajcency matrix of given network.}

\item{layers_def}{data.frame containing the modality ID, corresponding layer
in BN and maximal number of parents from given layer to GE nodes.}

\item{B_prior_mat}{a biological prior matrix.}

\item{omics}{named list containing the gene expression (possibly copy number
variation and methylation data). Each component of the list is a matrix
with samples in rows and features in columns.}
}
\value{
Numeric of length 1: neighborhood size
}
\description{
\code{neighborhood_size} This function is determines number of network structures
that can be reached from the current network structure.
}
\keyword{internal}
