% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omics}
\alias{omics}
\title{Omics data}
\format{
A named list with 3 components - each component corresponds to one omics data:
\describe{
each component of the list is a matrix with samples in rows and features in columns
}
}
\source{
\url{https://www.cancer.gov/tcga}
}
\usage{
omics
}
\description{
A named list containing the gene expression, copy number variation and methylation data.
}
\keyword{datasets}
