% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chain.R
\name{sample_chain}
\alias{sample_chain}
\title{Random initial network edge generation}
\usage{
sample_chain(empty_net, omics_ge)
}
\arguments{
\item{empty_net}{adjacency matrix of an empty network/graph
(all values are 0).}

\item{omics_ge}{matrix with gene expression data (samples in rows and
features in columns).}
}
\value{
BN object with conditional probabilities
}
\description{
\code{sample_chain} This function is used to sample random initial network.
The edges are sampled only between GE nodes.
}
\examples{
data(list=c("PK", "TFtarg_mat", "annot", "layers_def", "omics", 
"gene_annot"), package="IntOMICS")
omics <- omics_to_list(omics = omics, layers_def = layers_def, 
gene_annot = gene_annot)
B <- b_prior_mat(omics = omics, PK = PK, layers_def = layers_def, 
     annot = annot, lm_METH = TRUE, r_squared_thres = 0.3,
     p_val_thres = 0.05, TFtargs = TFtarg_mat, TFBS_belief = 0.75, 
     nonGE_belief = 0.5, woPKGE_belief = 0.5)
empty.net <- matrix(0, nrow = sum(mapply(ncol,B$omics)), ncol =
sum(mapply(ncol,B$omics)), dimnames = list(unlist(mapply(colnames,B$omics)),
unlist(mapply(colnames,B$omics))))
sample_chain(empty_net = empty.net, 
     omics_ge = B$omics[[layers_def$omics[1]]])
     
}
\keyword{internal}
