% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plots.R
\name{trace_plots}
\alias{trace_plots}
\title{Trace plots of MCMC simulation}
\usage{
trace_plots(mcmc_res, burn_in, thin, edge_freq_thres = NULL)
}
\arguments{
\item{mcmc_res}{MCMC_sapling_res output from the bn_module function.}

\item{burn_in}{numeric vector the minimal length of burn-in period
of the MCMC simulation.}

\item{thin}{numeric vector thinning frequency of the resulting
MCMC simulation.}

\item{edge_freq_thres}{numerical vector the quantile of all edge weights
used to filter the most reliable edges.}
}
\value{
MCMC simulation trace plots
}
\description{
\code{trace_plots} Create trace plots of MCMC simulation and filter low reliable
edges based on the edge_freq_thres parameter.
}
\examples{
if(interactive()){data("BN_mod_res", package="IntOMICS_R")
res_weighted <- trace_plots(mcmc_res = BN_mod_res, burn_in = 10000, 
       thin = 500, edge_freq_thres = 0.3)}

}
